reviewed algorithm run procedure + cleaned up unused files
