+++ src/config/submission-target.ts
@@ -15,3 +15,14 @@ export function computeProjectSubmissionTarget(
 
   return clamp(projectSubmissionTarget, [0, 12]);
 }
+
+export function adjustTarget(unstableTarget: number, targetModifier: number) {
+  return Math.max(unstableTarget + targetModifier, 0);
+}
+
+export function adjustUpperBound(
+  unstableUpperBound: number,
+  upperBoundModifier: number,
+) {
+  return Math.max(unstableUpperBound + upperBoundModifier, 0);
+}
